Option Screen Formatting Change, config updates
